// packages/lib/types/apiResponses.ts
export type OrderSide = 'Buy' | 'Sell' | 'Short' | 'Unknown';
export type OrderType =
  | 'Unknown'
  | 'Market order'
  | 'Limit'
  | 'StopMarket'
  | 'StopLimit'
  | 'TrailingStopMarket'
  | 'TrailingStopLimit'
  | 'BlockTrade';
export type OrderChangeReason =
  | 'Unknown'
  | 'NewInputAccepted'
  | 'NewInputRejected'
  | 'OtherRejected'
  | 'Expired'
  | 'Trade'
  | 'SystemCanceled_NoMoreMarket'
  | 'SystemCanceled_BelowMinimum'
  | 'SystemCanceled_PriceCollar'
  | 'SystemCanceled_MarginFailed'
  | 'UserModified';
export type OrderState =
  | 'Unknown'
  | 'Working'
  | 'Rejected'
  | 'Canceled'
  | 'Expired'
  | 'FullyExecuted';
// ...
export type DepositStatus =
  | 'New'
  | 'AdminProcessing'
  | 'Accepted'
  | 'Rejected'
  | 'SystemProcessing'
  | 'FullyProcessed'
  | 'Failed'
  | 'Pending'
  | 'Confirmed'
  | 'AmlProcessing'
  | 'AmlAccepted'
  | 'AmlRejected'
  | 'AmlFailed'
  | 'LimitsAccepted'
  | 'LimitsRejected'
  | 'AmlRegistered';

export type WithdrawStatus =
  | 'New'
  | 'AdminProcessing'
  | 'Accepted'
  | 'Rejected'
  | 'SystemProcessing'
  | 'FullyProcessed'
  | 'Failed'
  | 'Pending'
  | 'Pending2Fa'
  | 'AutoAccepted'
  | 'Delayed'
  | 'UserCancelled'
  | 'AdminCancelled'
  | 'AmlAddressVerificationProcessing'
  | 'AmlAccepted'
  | 'AmlRejected'
  | 'AmlFailed'
  | 'LimitsAccepted'
  | 'LimitsRejected'
  | 'Submitted'
  | 'Confirmed'
  | 'ManuallyConfirmed'
  | 'Confirmed2Fa'
  | 'AvsPending'
  | 'AmlAcceptedOverride'
  | 'PendingSubmission'
  | 'ReSubmitted'
  | 'ManualReview';

export type DepositTicket = {
  AssetManagerId: number;
  AccountProviderId: number;
  AccountId: number;
  AssetId: number;
  AccountName: string;
  AssetName: string;
  Amount: number;
  NotionalValue: number;
  NotionalProductId: number;
  OMSId: number;
  RequestCode: string;
  ReferenceId: string;
  RequestIP: string;
  RequestUser: number;
  RequestUserName: string;
  OperatorId: number;
  Status: DepositStatus;
  FeeAmt: number;
  UpdatedByUser: number;
  UpdatedByUserName: string;
  TicketNumber: number;
  DepositInfo: string;
  RejectReason?: string;
  CreatedTimestamp: string;
  LastUpdateTimeStamp: string;
  CreatedTimestampTick: number;
  LastUpdateTimeStampTick: number;
  Comments: any[];
  Attachments: any[];
};

export type WithdrawTicket = {
  AssetManagerId: number;
  AccountId: number;
  AccountName: string;
  AssetId: number;
  AssetName: string;
  Amount: number;
  NotionalValue: number;
  NotionalProductId: number;
  TemplateForm?: string;
  TemplateFormType?: string;
  OMSId: number;
  RequestCode: string;
  RequestIP: string;
  RequestUserId: number;
  RequestUserName: string;
  OperatorId: number;
  Status: WithdrawStatus;
  FeeAmt: number;
  UpdatedByUser: number;
  UpdatedByUserName: string;
  TicketNumber: number;
  WithdrawTransactionDetails: string;
  RejectReason: any;
  CreatedTimestamp: string;
  LastUpdateTimestamp: string;
  CreatedTimestampTick: number;
  LastUpdateTimestampTick: number;
  Comments: any[];
  Attachments: any[];
  AuditLog: any[];
};

export type TransferItem = {
  TransferId: number;
  SenderAccountId: number;
  SenderUserName: string;
  ReceiverAccountId: number;
  ReceiverUserName: string;
  ProductId: number;
  Amount: number;
  Notes: string;
  OMSId: number;
  Reason: string; // User
  TransferState: string; // Accepted
  CreatedTimeInTicks: number;
  LastUpdatedTimeInTicks: number;
};

export type AccountTradeItem = {
  OMSId: number;
  ExecutionId: number;
  TradeId: number;
  OrderId: number;
  AccountId: number;
  AccountName: string;
  SubAccountId: number;
  ClientOrderId: number;
  InstrumentId: number;
  Side: OrderSide;
  OrderType: OrderType;
  Quantity: number;
  RemainingQuantity: number;
  Price: number;
  Value: number;
  CounterParty: string;
  OrderTradeRevision: number;
  Direction: string; // NoChange
  IsBlockTrade: boolean;
  Fee: number;
  FeeProductId: number;
  OrderOriginator: number;
  UserName: string;
  TradeTimeMS: number;
  MakerTaker: string; // Maker
  AdapterTradeId: number;
  InsideBid: number;
  InsideBidSize: number;
  InsideAsk: number;
  InsideAskSize: number;
  IsQuote: boolean;
  CounterPartyClientUserId: number;
  NotionalProductId: number;
  NotionalRate: number;
  NotionalValue: number;
  NotionalHoldAmount: number;
  TradeTime: number;
};

export type WithdrawInfoResponse = Array<{
  AccountWithdrawInfoId: string;
  AccountId: number;
  ProductId: number;
  ProductName: string; // it's product full name
  Name: string;
  AccountIdentifier: string;
  Verified: boolean;
}>;

export type ActivityItem = {
  id: number;
  amount: number;
  type: string;
  status: string;
  timestamp: number;
  instrumentId?: number;
  productId?: number;
  decimalPlaces?: number;
  currency?: string;
  fullName?: string;
  symbol?: string;
};

export interface OrderActivity extends ActivityItem {
  type: OrderType;
  status: OrderState;
  iconSymbols?: readonly [string, string];
  side: OrderSide;
  avgPrice: number;
  limitPrice?: number;
}

export interface WalletActivity extends ActivityItem {
  type: 'withdraw' | 'deposit';
  status: DepositStatus | WithdrawStatus;
  iconSymbol?: string;
  notionalSymbol?: string;
  notionalRate?: number;
  notionalPlaces?: number;
  fullname?: string;
  comments?: string;
}

export type AccountTransactionType =
  | 'deposit'
  | 'withdraw'
  | 'buy'
  | 'sell'
  | 'send'
  | 'receive'
  | 'deposit-fee'
  | 'withdraw-fee'
  | 'buy-fee'
  | 'sell-fee'
  | 'unknown';

export interface AccountTransactionActivity extends ActivityItem {
  type: AccountTransactionType;
  status: 'complete';
  iconSymbol?: string;
  notionalSymbol?: string;
  notionalRate?: number;
  notionalPlaces?: number;
  referenceType: ReferenceType;
  transactionType: TransactionType;
  isCR: boolean;
}

export type TransactionType =
  | 'Fee'
  | 'Trade'
  | 'Other'
  | 'Reverse'
  | 'Hold'
  | 'Rebate'
  | 'MarginAcquisition'
  | 'MarginRelinquish';

export type ReferenceType =
  | 'Trade'
  | 'Deposit'
  | 'Withdraw'
  | 'Transfer'
  | 'OrderHold'
  | 'WithdrawHold'
  | 'DepositHold'
  | 'MarginHold'
  | 'ManualHold'
  | 'ManualEntry'
  | 'MarginAcquisition'
  | 'MarginRelinquish'
  | 'MarginQuoteHold';

export type AccountTransaction = {
  TransactionId: number;
  ReferenceId: number;
  OMSId: number;
  AccountId: number;
  CR: number;
  DR: number;
  Counterparty: number;
  TransactionType: TransactionType;
  ReferenceType: ReferenceType;
  ProductId: number;
  Balance: number;
  TimeStamp: number;
};